feat(scheduling): retire finance and stabilize thread-linked appointments

+++ app/services/access_control/legacy_custom_role_mapper.rb
@@ -114,7 +114,7 @@ class AccessControl::LegacyCustomRoleMapper
           grant.update!(access_scope: definition[:access_scope])
           changed = true
         end
-      else
+      elsif !grant.retired?
         grant.destroy!
         changed = true
       end
